refactor: project routes and tests for security and clarity
- Removed admin-specific project routes and tests.
- Consolidated project routes under a single handler with security decorators.
- Updated tests to mock user authentication and authorization.
- Improved error handling and validation in project routes.
- Cleaned up unused mocks and redundant code in test utilities.
